fix(decode): fifth-pass fuzz hardening — plane bounds + mcu_count width
Round 5 surfaced two more panics after the chroma-factor guard
(1c56eed):

- decode/pipeline.rs RGB-colorspace interleave: malformed sampling
  factors could leave component planes shorter than the Y plane,
  then the `for x in 0..out_width { r_row[x]; g_row[x]; b_row[x] }`
  loop OOB-panicked on the first row where the plane was short.
  Validate that every plane covers `(out_height-1)*stride + offset
  + out_width` before starting the interleave; surface CorruptData
  on mismatch instead of aborting.

- api/progressive_output.rs: `mcu_count: u16` overflows past 65535
  on large images with small restart_interval. Widen to u32 to
  match the baseline decode path (coefficient.rs uses u32) and
  cast `restart_interval as u32` at the comparison points.

fuzz_decompress_lenient now passes the round-4 seed; round-5 OOM
seeds (fuzz_read_coefficients, fuzz_transform) remain deferred to
a separate DoS-bounds pass.

All 167 test binaries still pass and testimages byte-exact
parity unchanged.

+++ src/api/progressive_output.rs
@@ -373,13 +373,13 @@ impl ProgressiveDecoder {
         if num_scan_components > 1 {
             // Interleaved scan (DC only in progressive)
             let mut dc_preds = [0i16; 4];
-            let mut mcu_count: u16 = 0;
+            let mut mcu_count: u32 = 0;
 
             for mcu_y in 0..self.mcus_y {
                 for mcu_x in 0..self.mcus_x {
                     if restart_interval > 0
                         && mcu_count > 0
-                        && mcu_count.is_multiple_of(restart_interval)
+                        && mcu_count.is_multiple_of(restart_interval as u32)
                     {
                         bit_reader.reset();
                         dc_preds = [0i16; 4];
@@ -438,7 +438,7 @@ impl ProgressiveDecoder {
             let blocks_y: usize = self.comp_infos[comp_idx].blocks_y;
             let mut dc_pred: i16 = 0;
             let mut eob_run: u16 = 0;
-            let mut mcu_count: u16 = 0;
+            let mut mcu_count: u32 = 0;
 
             let dc_table_ref = if is_dc {
                 Some(
@@ -474,7 +474,7 @@ impl ProgressiveDecoder {
                 for bx in 0..blocks_x {
                     if restart_interval > 0
                         && mcu_count > 0
-                        && mcu_count.is_multiple_of(restart_interval)
+                        && mcu_count.is_multiple_of(restart_interval as u32)
                     {
                         bit_reader.reset();
                         dc_pred = 0;
